+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.2
+ - Updated to work with latest version of the game (Feb 10th, 2026)
+ - Removed Quota Modifier setting in favor of Soft Compatibility with [QuotaQueen](https://thunderstore.io/c/yapyap/p/Robyn/QuotaQueen/)
+	- See readme for details
+
 ## 0.1.1
  - Updated changelog from the mod base template, lol.
  - Added new setting Quota Modifier, which allows you to modify the quota based on the number of players in the lobby.